+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Bug fix:
 - [#985](https://github.com/elysiajs/elysia/issues/924) Nullable accept options
 - [#1028](https://github.com/elysiajs/elysia/issues/1028) string | string[] query parameter, reference array
 - [#1120](https://github.com/elysiajs/elysia/issues/1120) cannot set multiple cookies when response is a file
+- [#1124](https://github.com/elysiajs/elysia/issues/1124) validate url encoded query
 - [#1274](https://github.com/elysiajs/elysia/issues/1274) support .use(undefined | false) for conditional plugin
 - [#1276](https://github.com/elysiajs/elysia/issues/1276) mapResponse with set inference produce invalid instruction
 - [#1268](https://github.com/elysiajs/elysia/issues/1268) using number instead of stringifed value for reporting validation error